+++ pybamm/expression_tree/operations/evaluate.py
@@ -17,6 +17,98 @@ if system() != "Windows":
 
     config.update("jax_enable_x64", True)
 
+    class JaxCooMatrix:
+        """
+        A sparse matrix in COO format, with internal arrays using jax device arrays
+
+        This matrix only has two operations supported, a multiply with a scalar, and a
+        dot product with a dense vector. It can also be converted to a dense 2D jax
+        device array
+
+        Parameters
+        ----------
+
+        row: arraylike
+            1D array holding row indices of non-zero entries
+        col: arraylike
+            1D array holding col indices of non-zero entries
+        data: arraylike
+            1D array holding non-zero entries
+        shape: 2-element tuple (x, y)
+            where x is the number of rows, and y the number of columns of the matrix
+        """
+
+        def __init__(self, row, col, data, shape):
+            self.row = jax.numpy.array(row)
+            self.col = jax.numpy.array(col)
+            self.data = jax.numpy.array(data)
+            self.shape = shape
+            self.nnz = len(self.data)
+
+        def toarray(self):
+            """convert sparse matrix to a dense 2D array"""
+            result = jax.numpy.zeros(self.shape, dtype=self.data.dtype)
+            return result.at[self.row, self.col].add(self.data)
+
+        def dot_product(self, b):
+            """
+            dot product of matrix with a dense column vector b
+
+            Parameters
+            ----------
+            b: jax device array
+                must have shape (n, 1)
+            """
+            # assume b is a column vector
+            result = jax.numpy.zeros((self.shape[0], 1), dtype=b.dtype)
+            return result.at[self.row].add(self.data.reshape(-1, 1) * b[self.col])
+
+        def scalar_multiply(self, b):
+            """
+            multiply of matrix with a scalar b
+
+            Parameters
+            ----------
+            b: Number or 1 element jax device array
+                scalar value to multiply
+            """
+            # assume b is a scalar or ndarray with 1 element
+            return JaxCooMatrix(
+                self.row, self.col,
+                (self.data * b).reshape(-1),
+                self.shape
+            )
+
+        def multiply(self, b):
+            """
+            general matrix multiply not supported
+            """
+            raise NotImplementedError
+
+        def __matmul__(self, b):
+            """see self.dot_product"""
+            return self.dot_product(b)
+
+    def create_jax_coo_matrix(value):
+        """
+        Creates a JaxCooMatrix from a scipy.sparse matrix
+
+        Parameters
+        ----------
+
+        value: scipy.sparse matrix
+            the sparse matrix to be converted
+        """
+        scipy_coo = value.tocoo()
+        row = jax.numpy.asarray(scipy_coo.row)
+        col = jax.numpy.asarray(scipy_coo.col)
+        data = jax.numpy.asarray(scipy_coo.data)
+        return JaxCooMatrix(row, col, data, value.shape)
+else:
+
+    def create_jax_coo_matrix(value):  # pragma: no cover
+        raise NotImplementedError('Jax is not available on Windows')
+
 
 def id_to_python_variable(symbol_id, constant=False):
     """
@@ -33,7 +125,15 @@ def id_to_python_variable(symbol_id, constant=False):
     return var_format.format(symbol_id).replace("-", "m")
 
 
-def find_symbols(symbol, constant_symbols, variable_symbols, to_dense=False):
+def is_scalar(arg):
+    is_number = isinstance(arg, numbers.Number)
+    if is_number:
+        return True
+    else:
+        return np.all(np.array(arg.shape) == 1)
+
+
+def find_symbols(symbol, constant_symbols, variable_symbols, output_jax=False):
     """
     This function converts an expression tree to a dictionary of node id's and strings
     specifying valid python code to calculate that nodes value, given y and t.
@@ -44,7 +144,7 @@ def find_symbols(symbol, constant_symbols, variable_symbols, to_dense=False):
     `variable_symbols`
 
     Note that it is important that the arguments `constant_symbols` and
-    `variable_symbols` be and *ordered* dict, since the final ordering of the code lines
+    `variable_symbols` be an *ordered* dict, since the final ordering of the code lines
     are important for the calculations. A dict is specified rather than a list so that
     identical subtrees (which give identical id's) are not recalculated in the code
 
@@ -59,22 +159,29 @@ def find_symbols(symbol, constant_symbols, variable_symbols, to_dense=False):
     variable_symbol: collections.OrderedDict
         The output dictionary of variable (with y or t) symbol ids to lines of code
 
-    to_dense: bool
-        If True, all constants and expressions are converted to using dense matrices
+    output_jax: bool
+        If True, only numpy and jax operations will be used in the generated code,
+        raises NotImplNotImplementedError if any SparseStack or Mat-Mat multiply
+        operations are used
 
     """
+    # constant symbols that are not numbers are stored in a list of constants, which are
+    # passed into the generated function constant symbols that are numbers are written
+    # directly into the code
     if symbol.is_constant():
         value = symbol.evaluate()
         if not isinstance(value, numbers.Number):
-            if to_dense and scipy.sparse.issparse(value):
-                constant_symbols[symbol.id] = value.toarray()
+            if output_jax and scipy.sparse.issparse(value):
+                # convert any remaining sparse matrices to our custom coo matrix
+                constant_symbols[symbol.id] = create_jax_coo_matrix(value)
+
             else:
                 constant_symbols[symbol.id] = value
         return
 
     # process children recursively
     for child in symbol.children:
-        find_symbols(child, constant_symbols, variable_symbols, to_dense)
+        find_symbols(child, constant_symbols, variable_symbols, output_jax)
 
     # calculate the variable names that will hold the result of calculating the
     # children variables
@@ -96,36 +203,52 @@ def find_symbols(symbol, constant_symbols, variable_symbols, to_dense=False):
         if isinstance(symbol, pybamm.Multiplication):
             dummy_eval_left = symbol.children[0].evaluate_for_shape()
             dummy_eval_right = symbol.children[1].evaluate_for_shape()
-            if not to_dense and scipy.sparse.issparse(dummy_eval_left):
-                symbol_str = "{0}.multiply({1})".format(
-                    children_vars[0], children_vars[1]
-                )
-            elif not to_dense and scipy.sparse.issparse(dummy_eval_right):
-                symbol_str = "{1}.multiply({0})".format(
-                    children_vars[0], children_vars[1]
-                )
+            if scipy.sparse.issparse(dummy_eval_left):
+                if output_jax and is_scalar(dummy_eval_right):
+                    symbol_str = "{0}.scalar_multiply({1})"\
+                        .format(children_vars[0], children_vars[1])
+                else:
+                    symbol_str = "{0}.multiply({1})"\
+                        .format(children_vars[0], children_vars[1])
+            elif scipy.sparse.issparse(dummy_eval_right):
+                if output_jax and is_scalar(dummy_eval_left):
+                    symbol_str = "{1}.scalar_multiply({0})"\
+                        .format(children_vars[0], children_vars[1])
+                else:
+                    symbol_str = "{1}.multiply({0})"\
+                        .format(children_vars[0], children_vars[1])
             else:
                 symbol_str = "{0} * {1}".format(children_vars[0], children_vars[1])
         elif isinstance(symbol, pybamm.Division):
             dummy_eval_left = symbol.children[0].evaluate_for_shape()
-            if not to_dense and scipy.sparse.issparse(dummy_eval_left):
-                symbol_str = "{0}.multiply(1/{1})".format(
-                    children_vars[0], children_vars[1]
-                )
+            dummy_eval_right = symbol.children[1].evaluate_for_shape()
+            if scipy.sparse.issparse(dummy_eval_left):
+                if output_jax and is_scalar(dummy_eval_right):
+                    symbol_str = "{0}.scalar_multiply(1/{1})"\
+                        .format(children_vars[0], children_vars[1])
+                else:
+                    symbol_str = "{0}.multiply(1/{1})"\
+                        .format(children_vars[0], children_vars[1])
             else:
                 symbol_str = "{0} / {1}".format(children_vars[0], children_vars[1])
 
         elif isinstance(symbol, pybamm.Inner):
             dummy_eval_left = symbol.children[0].evaluate_for_shape()
             dummy_eval_right = symbol.children[1].evaluate_for_shape()
-            if not to_dense and scipy.sparse.issparse(dummy_eval_left):
-                symbol_str = "{0}.multiply({1})".format(
-                    children_vars[0], children_vars[1]
-                )
-            elif not to_dense and scipy.sparse.issparse(dummy_eval_right):
-                symbol_str = "{1}.multiply({0})".format(
-                    children_vars[0], children_vars[1]
-                )
+            if scipy.sparse.issparse(dummy_eval_left):
+                if output_jax and is_scalar(dummy_eval_right):
+                    symbol_str = "{0}.scalar_multiply({1})"\
+                        .format(children_vars[0], children_vars[1])
+                else:
+                    symbol_str = "{0}.multiply({1})"\
+                        .format(children_vars[0], children_vars[1])
+            elif scipy.sparse.issparse(dummy_eval_right):
+                if output_jax and is_scalar(dummy_eval_left):
+                    symbol_str = "{1}.scalar_multiply({0})"\
+                        .format(children_vars[0], children_vars[1])
+                else:
+                    symbol_str = "{1}.multiply({0})"\
+                        .format(children_vars[0], children_vars[1])
             else:
                 symbol_str = "{0} * {1}".format(children_vars[0], children_vars[1])
 
@@ -133,6 +256,19 @@ def find_symbols(symbol, constant_symbols, variable_symbols, to_dense=False):
             symbol_str = "np.minimum({},{})".format(children_vars[0], children_vars[1])
         elif isinstance(symbol, pybamm.Maximum):
             symbol_str = "np.maximum({},{})".format(children_vars[0], children_vars[1])
+
+        elif isinstance(symbol, pybamm.MatrixMultiplication):
+            dummy_eval_left = symbol.children[0].evaluate_for_shape()
+            dummy_eval_right = symbol.children[1].evaluate_for_shape()
+            if output_jax and (
+                    scipy.sparse.issparse(dummy_eval_left) and
+                    scipy.sparse.issparse(dummy_eval_right)
+            ):
+                raise NotImplementedError('sparse mat-mat multiplication not supported '
+                                          'for output_jax == True')
+            else:
+                symbol_str = children_vars[0] + " " + symbol.name + " " \
+                    + children_vars[1]
         else:
             symbol_str = children_vars[0] + " " + symbol.name + " " + children_vars[1]
 
@@ -172,10 +308,12 @@ def find_symbols(symbol, constant_symbols, variable_symbols, to_dense=False):
                 symbol_str = "{}".format(",".join(children_vars))
 
         elif isinstance(symbol, pybamm.SparseStack):
-            if not to_dense and len(children_vars) > 1:
-                symbol_str = "scipy.sparse.vstack(({}))".format(",".join(children_vars))
-            elif len(children_vars) > 1:
-                symbol_str = "np.vstack(({}))".format(",".join(children_vars))
+            if len(children_vars) > 1:
+                if output_jax:
+                    raise NotImplementedError
+                else:
+                    symbol_str = "scipy.sparse.vstack(({}))".format(
+                        ",".join(children_vars))
             else:
                 symbol_str = "{}".format(",".join(children_vars))
 
@@ -230,7 +368,7 @@ def find_symbols(symbol, constant_symbols, variable_symbols, to_dense=False):
     variable_symbols[symbol.id] = symbol_str
 
 
-def to_python(symbol, debug=False, to_dense=False):
+def to_python(symbol, debug=False, output_jax=False):
     """
     This function converts an expression tree into a dict of constant input values, and
     valid python code that acts like the tree's :func:`pybamm.Symbol.evaluate` function
@@ -250,14 +388,15 @@ def to_python(symbol, debug=False, to_dense=False):
         the expression tree
     str:
         valid python code that will evaluate all the variable nodes in the tree.
-    to_dense: bool
-        If True, all constants and expressions are converted to using dense matrices
+    output_jax: bool
+        If True, only numpy and jax operations will be used in the generated code.
+        Raises NotImplNotImplementedError if any SparseStack or Mat-Mat multiply
+        operations are used
 
     """
-
     constant_values = OrderedDict()
     variable_symbols = OrderedDict()
-    find_symbols(symbol, constant_values, variable_symbols, to_dense)
+    find_symbols(symbol, constant_values, variable_symbols, output_jax)
 
     line_format = "{} = {}"
 
@@ -267,7 +406,7 @@ def to_python(symbol, debug=False, to_dense=False):
                 line_format.format(id_to_python_variable(symbol_id, False), symbol_line)
             )
             + line_format.format(id_to_python_variable(symbol_id, False), symbol_line)
-            + "; print(type({0}),{0}.shape)".format(
+            + "; print(type({0}),np.shape({0}))".format(
                 id_to_python_variable(symbol_id, False)
             )
             for symbol_id, symbol_line in variable_symbols.items()
@@ -365,11 +504,6 @@ class EvaluatorJax:
     so any sparse matrices and operations involved sparse matrices are converted to
     their dense equivilents before compilation
 
-    Raises
-    ------
-    RuntimeError
-        if any sparse matrices are present in the expression tree
-
     Parameters
     ----------
 
@@ -380,7 +514,7 @@ class EvaluatorJax:
     """
 
     def __init__(self, symbol):
-        constants, python_str = pybamm.to_python(symbol, debug=False, to_dense=True)
+        constants, python_str = pybamm.to_python(symbol, debug=False, output_jax=True)
 
         # replace numpy function calls to jax numpy calls
         python_str = python_str.replace("np.", "jax.numpy.")
@@ -390,22 +524,34 @@ class EvaluatorJax:
             if isinstance(constants[symbol_id], np.ndarray):
                 constants[symbol_id] = jax.device_put(constants[symbol_id])
 
-        # extract constants in generated function
-        for i, symbol_id in enumerate(constants.keys()):
-            const_name = id_to_python_variable(symbol_id, True)
-            python_str = "{} = constants[{}]\n".format(const_name, i) + python_str
+        # get a list of constant arguments to input to the function
+        arg_list = [
+            id_to_python_variable(symbol_id, True)
+            for symbol_id in constants.keys()
+        ]
 
-        # constants passed in as an ordered dict, convert to list
-        self._constants = list(constants.values())
+        # get a list of hashable arguments to make static
+        # a jax device array is not hashable
+        static_argnums = (
+            i for i, c in enumerate(constants.values())
+            if not (
+                isinstance(c, jax.interpreters.xla.DeviceArray)
+            )
+        )
+
+        # store constants
+        self._constants = tuple(constants.values())
 
         # indent code
         python_str = "   " + python_str
         python_str = python_str.replace("\n", "\n   ")
 
         # add function def to first line
+        args = 't=None, y=None, y_dot=None, inputs=None, known_evals=None'
+        if arg_list:
+            args = ','.join(arg_list) + ', ' + args
         python_str = (
-            "def evaluate_jax(constants, t=None, y=None, "
-            "y_dot=None, inputs=None, known_evals=None):\n" + python_str
+            "def evaluate_jax({}):\n".format(args) + python_str
         )
 
         # calculate the final variable that will output the result of calling `evaluate`
@@ -423,15 +569,22 @@ class EvaluatorJax:
         # store a copy of examine_jaxpr
         python_str = python_str + "\nself._evaluate_jax = evaluate_jax"
 
+        # store the final generated code
+        self._python_str = python_str
+
         # compile and run the generated python code,
         compiled_function = compile(python_str, result_var, "exec")
         exec(compiled_function)
 
-        self._jit_evaluate = jax.jit(self._evaluate_jax, static_argnums=(0, 4, 5))
+        n = len(arg_list)
+        static_argnums = tuple(static_argnums)
+        self._jit_evaluate = jax.jit(self._evaluate_jax,
+                                     static_argnums=static_argnums)
 
         # store a jit version of evaluate_jax's jacobian
-        jacobian_evaluate = jax.jacfwd(self._evaluate_jax, argnums=2)
-        self._jac_evaluate = jax.jit(jacobian_evaluate, static_argnums=(0, 4, 5))
+        jacobian_evaluate = jax.jacfwd(self._evaluate_jax, argnums=1 + n)
+        self._jac_evaluate = jax.jit(jacobian_evaluate,
+                                     static_argnums=static_argnums)
 
     def get_jacobian(self):
         return EvaluatorJaxJacobian(self._jac_evaluate, self._constants)
@@ -443,7 +596,7 @@ class EvaluatorJax:
 
         # execute code
         jaxpr = jax.make_jaxpr(self._evaluate_jax)(
-            self._constants, t, y, y_dot, inputs, known_evals
+            *self._constants, t, y, y_dot, inputs, known_evals
         ).jaxpr
         print("invars:", jaxpr.invars)
         print("outvars:", jaxpr.outvars)
@@ -461,7 +614,7 @@ class EvaluatorJax:
         if y is not None and y.ndim == 1:
             y = y.reshape(-1, 1)
 
-        result = self._jit_evaluate(self._constants, t, y, y_dot, inputs, known_evals)
+        result = self._jit_evaluate(*self._constants, t, y, y_dot, inputs, known_evals)
 
         # don't need known_evals, but need to reproduce Symbol.evaluate signature
         if known_evals is not None:
@@ -484,7 +637,7 @@ class EvaluatorJaxJacobian:
             y = y.reshape(-1, 1)
 
         # execute code
-        result = self._jac_evaluate(self._constants, t, y, y_dot, inputs, known_evals)
+        result = self._jac_evaluate(*self._constants, t, y, y_dot, inputs, known_evals)
         result = result.reshape(result.shape[0], -1)
 
         # don't need known_evals, but need to reproduce Symbol.evaluate signature