+++ pybamm/solvers/jax_solver.py
@@ -60,6 +60,7 @@ class JaxSolver(pybamm.BaseSolver):
         self.extra_options = extra_options or {}
         self.name = "JAX solver ({})".format(method)
         self._cached_solves = dict()
+        pybamm.citations.register("jax2018github")
 
     def get_solve(self, model, t_eval):
         """
@@ -192,10 +193,10 @@ class JaxSolver(pybamm.BaseSolver):
         if model not in self._cached_solves:
             self._cached_solves[model] = self.create_solve(model, t_eval)
 
-        y = self._cached_solves[model](inputs)
+        y = self._cached_solves[model](inputs).block_until_ready()
         integration_time = timer.time()
 
-        # note - the actual solve is not done until this line!
+        # convert to a normal numpy array
         y = onp.array(y)
 
         termination = "final time"