+++ tests/unit/test_expression_tree/test_unary_operators.py
@@ -447,7 +447,6 @@ class TestUnaryOperators(unittest.TestCase):
             ["negative electrode"],
             ["separator"],
             ["positive electrode"],
-            ["negative electrode", "separator", "positive electrode"],
         ]:
             a = pybamm.Symbol("a", domain=domain)
             x = pybamm.SpatialVariable("x", domain)
@@ -457,7 +456,16 @@ class TestUnaryOperators(unittest.TestCase):
             self.assertEqual(av_a.children[0].integration_variable[0].domain, x.domain)
             self.assertEqual(av_a.domain, [])
 
-        a = pybamm.Symbol("a", domain="new domain")
+        # whole electrode domain is different as the division by 1 gets simplified out
+        domain = ["negative electrode", "separator", "positive electrode"]
+        a = pybamm.Variable("a", domain=domain)
+        x = pybamm.SpatialVariable("x", domain)
+        av_a = pybamm.x_average(a)
+        self.assertIsInstance(av_a, pybamm.Integral)
+        self.assertEqual(av_a.integration_variable[0].domain, x.domain)
+        self.assertEqual(av_a.domain, [])
+
+        a = pybamm.Variable("a", domain="new domain")
         av_a = pybamm.x_average(a)
         self.assertEqual(av_a.domain, [])
         self.assertIsInstance(av_a, pybamm.Division)
@@ -581,6 +589,23 @@ class TestUnaryOperators(unittest.TestCase):
             ValueError, "Can't take the z-average of a symbol that evaluates on edges"
         ):
             pybamm.z_average(symbol_on_edges)
+
+    def test_unary_simplifications(self):
+        a = pybamm.Scalar(0, domain="domain")
+        b = pybamm.Scalar(1)
+        d = pybamm.Scalar(-1)
+
+        # negate
+        self.assertIsInstance((-a), pybamm.Scalar)
+        self.assertEqual((-a).evaluate(), 0)
+        self.assertIsInstance((-b), pybamm.Scalar)
+        self.assertEqual((-b).evaluate(), -1)
+
+        # absolute value
+        self.assertIsInstance((abs(a)), pybamm.Scalar)
+        self.assertEqual((abs(a)).evaluate(), 0)
+        self.assertIsInstance((abs(d)), pybamm.Scalar)
+        self.assertEqual((abs(d)).evaluate(), 1)
 
 
 if __name__ == "__main__":