+++ tests/unit/test_models/test_full_battery_models/test_lithium_ion/test_spme.py
@@ -110,13 +110,13 @@ class TestSPMe(unittest.TestCase):
 
     def test_surface_form_differential(self):
         options = {"surface form": "differential"}
-        with self.assertRaises(NotImplementedError):
-            pybamm.lithium_ion.SPMe(options)
+        model = pybamm.lithium_ion.SPMe(options)
+        model.check_well_posedness()
 
     def test_surface_form_algebraic(self):
         options = {"surface form": "algebraic"}
-        with self.assertRaises(NotImplementedError):
-            pybamm.lithium_ion.SPMe(options)
+        model = pybamm.lithium_ion.SPMe(options)
+        model.check_well_posedness()
 
     def test_integrated_conductivity(self):
         options = {"electrolyte conductivity": "integrated"}
@@ -152,6 +152,33 @@ class TestSPMeWithSEI(unittest.TestCase):
 
     def test_well_posed_ec_reaction_limited(self):
         options = {"sei": "ec reaction limited", "sei porosity change": True}
+        model = pybamm.lithium_ion.SPMe(options)
+        model.check_well_posedness()
+
+
+class TestSPMeWithCrack(unittest.TestCase):
+    def test_well_posed_none_crack(self):
+        options = {"particle": "Fickian diffusion", "particle cracking": None}
+        model = pybamm.lithium_ion.SPMe(options)
+        model.check_well_posedness()
+
+    def test_well_posed_no_cracking(self):
+        options = {"particle": "Fickian diffusion", "particle cracking": "no cracking"}
+        model = pybamm.lithium_ion.SPMe(options)
+        model.check_well_posedness()
+
+    def test_well_posed_anode_cracking(self):
+        options = {"particle": "Fickian diffusion", "particle cracking": "anode"}
+        model = pybamm.lithium_ion.SPMe(options)
+        model.check_well_posedness()
+
+    def test_well_posed_cathode_cracking(self):
+        options = {"particle": "Fickian diffusion", "particle cracking": "cathode"}
+        model = pybamm.lithium_ion.SPMe(options)
+        model.check_well_posedness()
+
+    def test_well_posed_both_cracking(self):
+        options = {"particle": "Fickian diffusion", "particle cracking": "both"}
         model = pybamm.lithium_ion.SPMe(options)
         model.check_well_posedness()